refactor: reorganize address type imports and implement form validation schema Signed-off-by: Justus Kleinau <justus@kleinau-online.de>
